asciidoclog: add extra option to allow appending
Some functions may need to append to an existing file. Provide an
option to allow that.

If not specified, the option defaults to overwrite instead of
append. This preserves the old behavior for existing code.

+++ elbepack/asciidoclog.py
@@ -89,11 +89,14 @@ class LogBase(object):
 
 
 class ASCIIDocLog (LogBase):
-    def __init__(self, fname):
+    def __init__(self, fname, append=False):
         self.fname = fname
-        if os.path.isfile(fname):
-            os.unlink(fname)
-        fp = file(fname, "w", 0)
+        if append:
+            fp = file(fname, "a", 0)
+        else:
+            if os.path.isfile(fname):
+                os.unlink(fname)
+            fp = file(fname, "w", 0)
 
         LogBase.__init__(self, fp)
